+++ javaScript/projects/weather-app/src/components/home/home.js
@@ -1,4 +1,4 @@
-import createElement from '../../utilities/createElement';
+import createElement from '../../helpers/createElement';
 
 const homeBuilder = {
   cacheDOM() {