+++ javaScript/projects/weather-app/src/helpers/importAll.js
@@ -1,0 +1,10 @@
+export default function importAll(r) {
+  const files = {};
+  const filesArr = [];
+  r.keys().forEach((item) => {
+    files[item.replace('./', '')] = r(item);
+    filesArr.push(item.replace('./', ''));
+  });
+
+  return { files, filesArr };
+}